# contracts/openapi/control-plane-v1.yaml
openapi: 3.1.0
info:
  title: Molejo Control Plane API
  version: 1.0.0
  description: Pre-alpha control-plane contract for the local beta workspace.
servers:
  - url: /
paths:
...
  /api/v1/workspaces/{workspaceId}/projects/{projectId}/apps/{appId}/environments/{appEnvironmentId}/observability/logs:
    parameters:
      - { $ref: '#/components/parameters/WorkspaceId' }
      - { $ref: '#/components/parameters/ProjectId' }
      - { $ref: '#/components/parameters/AppId' }
      - { $ref: '#/components/parameters/AppEnvironmentId' }
    get:
      operationId: listAppEnvironmentRuntimeLogs
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: search, in: query, schema: { type: string, maxLength: 200 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 1000, default: 200 } }
        - { name: cursor, in: query, description: Opaque keyset cursor returned by the previous page, schema: { type: string, maxLength: 1024 } }
      responses:
        '200':
          description: Bounded sanitized runtime logs
          content: { application/json: { schema: { $ref: '#/components/schemas/RuntimeLogs' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /api/v1/workspaces/{workspaceId}/projects/{projectId}/apps/{appId}/environments/{appEnvironmentId}/observability/events:
    parameters:
      - { $ref: '#/components/parameters/WorkspaceId' }
      - { $ref: '#/components/parameters/ProjectId' }
      - { $ref: '#/components/parameters/AppId' }
      - { $ref: '#/components/parameters/AppEnvironmentId' }
    get:
      operationId: listAppEnvironmentRuntimeEvents
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
      responses:
        '200':
          description: Durable control-plane and Kubernetes runtime events
          content: { application/json: { schema: { $ref: '#/components/schemas/RuntimeEvents' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: fruto_session
  parameters:
...
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string, pattern: '^ws-[a-z2-7]{20}$' }
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, pattern: '^prj-[a-z2-7]{20}$' }
...
    AppId:
      name: appId
      in: path
      required: true
      schema: { type: string, pattern: '^app-[a-z2-7]{20}$' }
    AppEnvironmentId:
      name: appEnvironmentId
      in: path
      required: true
      schema: { type: string, pattern: '^aev-[a-z2-7]{20}$' }
...
  responses:
...
    BadRequest:
      description: Invalid request
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unavailable:
      description: External integration is not configured or unavailable
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
...
    RuntimeLog:
      type: object
      required: [id, timestamp, body, severity]
      additionalProperties: false
      properties:
        id: { type: string, pattern: '^log-[a-f0-9]{32}$' }
        timestamp: { type: string, format: date-time }
        body: { type: string, maxLength: 65536 }
        severity: { type: string, maxLength: 253 }
    RuntimeLogs:
      type: object
      required: [from, to, liveCursor, items]
      additionalProperties: false
      properties:
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        liveCursor: { type: string }
        nextCursor: { type: [string, 'null'] }
        items: { type: array, items: { $ref: '#/components/schemas/RuntimeLog' } }
...
    RuntimeEvent:
      type: object
      required: [timestamp, source, type, reason, message]
      additionalProperties: false
      properties:
        timestamp: { type: string, format: date-time }
        source: { type: string, enum: [control-plane, kubernetes] }
        type: { type: string, maxLength: 253 }
        reason: { type: string, maxLength: 253 }
        message: { type: string, maxLength: 65536 }
    RuntimeEvents:
      type: object
      required: [from, to, items]
      additionalProperties: false
      properties:
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        items: { type: array, items: { $ref: '#/components/schemas/RuntimeEvent' } }
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }